fix: moved bin prefix logic to new file
